Enable delete and add cohort placeholders
