+++ apps/vue-mri-ui-lib/src/utils/BookmarkUtils.ts
@@ -11,7 +11,7 @@ export function formatBookmark(bookmark: FormattedBookmark) {
   }
 
   const filterCards = bookmarkObj.filter.cards
-  const boolContainers = filterCards.content
+  const filterCardsContent = filterCards.content
 
   return {
     id: bookmark.bmkId,
@@ -22,11 +22,10 @@ export function formatBookmark(bookmark: FormattedBookmark) {
     version: bookmark.version,
     dateModified: DateUtils.displayBookmarkDateFormat(bookmark.modified),
     timeModified: DateUtils.displayBookmarkTimeFormat(bookmark.modified),
-    filterCardData: boolContainers,
+    filterCardData: filterCardsContent,
     chartType: bookmarkObj.chartType,
     axisInfo: bookmarkObj.chartType === 'list' ? bookmarkObj.filter.selected_attributes : bookmarkObj.axisSelection,
     shared: bookmark.shared,
-    __type: 'D2E_COHORT_DEFINITION',
   }
 }
 
@@ -39,16 +38,50 @@ export function formatAtlasCohortDefinition(atlasCD: FormattedAtlasCohortDefinit
     ...atlasCD,
     createdOn: DateUtils.displayBookmarkDateFormat(atlasCD.createdOn),
     updatedOn: DateUtils.displayBookmarkDateFormat(atlasCD.updatedOn),
-    __type: 'ATLAS_COHORT_DEFINITION',
   }
 }
 
 export function formatCohortDefinition(cohortDefinition: FormattedMaterializedCohort) {
   return {
-    __type: 'MATERIALIZED_COHORT',
     id: cohortDefinition.id,
     patientCount: cohortDefinition.patientCount,
     cohortDefinitionName: cohortDefinition.cohortDefinitionName,
     createdOn: DateUtils.displayBookmarkDateFormat(cohortDefinition.createdOn),
+  }
+}
+
+/**
+ * Determines the type of bookmark based on the properties of the BookmarkDisplay object.
+ *
+ * @param {BookmarkDisplay} obj - The BookmarkDisplay object to analyze.
+ * @returns {'A' | 'D' | 'M' | 'A+M' | 'D+M'} The type of bookmark:
+ *   - 'A': Atlas Cohort Definition
+ *   - 'D': D2E Cohort Definition
+ *   - 'M': Materialized Cohort
+ *   - 'A+M': Atlas Cohort Definition + Materialized Cohort
+ *   - 'D+M': D2E Cohort Definition + Materialized Cohort
+ 
+ * @example
+ * const bookmark = {
+ *   cohortDefinition: true,
+ *   atlasCohortDefinition: true
+ * };
+ * const type = getBookmarkType(bookmark); // Returns 'A+M'
+ */
+export function getBookmarkType(obj: BookmarkDisplay): BookmarkType {
+  if (obj.cohortDefinition) {
+    if (obj.atlasCohortDefinition) {
+      return 'A+M'
+    }
+    if (obj.bookmark) {
+      return 'D+M'
+    }
+    return 'M'
+  }
+  if (obj.atlasCohortDefinition) {
+    return 'A'
+  }
+  if (obj.bookmark) {
+    return 'D'
   }
 }
